requirements.txt
Flask==1.1.2
github3.py==1.3.0
GitPython==3.1.3